server/requirements.txt
#Chatbot  
flask==3.0.0
flask-cors==4.0.0
google-generativeai>=0.7.0
python-dotenv>=1.0.0

# Voice Model 
torch
transformers
openai-whisper
numpy
sentencepiece